+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KF392YKGA6KMWGJS7ZGN6GZQ"
+        "01KF5NV6EQRV4S51TS1J514NNR"
     )
     r"""The unique identifier of the data source"""