+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -96,7 +96,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-01-16T11:29:31.996Z")
+    created: Optional[datetime] = parse_datetime("2026-01-17T09:50:57.877Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")