+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -231,7 +231,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kf392y6rvwp4pn8391y41bpp"
+        "conv_01kf5nv5zc6nbv89x50n0gm80z"
     )
     r"""Unique conversation identifier with `conv_` prefix."""