+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatapointop.py
@@ -1445,7 +1445,7 @@ class UpdateDatapointEvaluations3(BaseModel):
 
     source: Optional[UpdateDatapointEvaluationsDatasetsResponseSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-01-16T11:29:41.647Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-01-17T09:51:07.919Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1513,7 +1513,7 @@ class UpdateDatapointEvaluations2(BaseModel):
 
     source: Optional[UpdateDatapointEvaluationsDatasetsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-01-16T11:29:41.647Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-01-17T09:51:07.919Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1581,7 +1581,7 @@ class UpdateDatapointEvaluations1(BaseModel):
 
     source: Optional[UpdateDatapointEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-01-16T11:29:41.646Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-01-17T09:51:07.918Z")
     r"""The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1684,7 +1684,7 @@ class UpdateDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-16T11:29:28.506Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-17T09:50:54.473Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")